Post a boolean to SmartDashboard if the bot is aligned with the chain
To help align the bot with the angle of the chain without being
able to see the April Tags, this uses the gyro angle to provide a
quick reference for whether the bot is aligned with one of the 3
angles. On SmartDashboard, the boolean field can be changed to
display a simple red or green box based on the boolean input value.